requirements.txt
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.17.0
altair==4.2.2
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asgiref==3.10.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==4.0.3
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.13.5
bitsandbytes==0.48.2
bleach==6.2.0
blinker==1.9.0
blis==1.3.0
boto3==1.40.60
botocore==1.40.60
build==1.3.0
cachetools==6.2.0
catalogue==2.0.10
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.3.0
cloudpathlib==0.23.0
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
comm==0.2.3
confection==0.1.5
contourpy==1.3.2
cryptography==46.0.3
cut-cross-entropy==25.1.1
cx-Oracle==8.3.0
cycler==0.12.1
cymem==2.0.11
dacite==1.6.0
dagshub==0.6.3
dagshub-annotation-converter==0.1.15
databases==0.9.0
databricks-sdk==0.70.0
dataclasses-json==0.6.7
datasets==4.3.0
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.3.1
diffusers==0.35.2
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.17.0
durationpy==0.10
email-validator==2.3.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
entrypoints==0.4
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.115.6
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.0
fastjsonschema==2.21.2
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
googleapis-common-protos==1.71.0
gql==4.0.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
hf_transfer==0.1.9
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.10
imbalanced-learn==0.14.0
importlib_metadata==8.4.0
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.30.1
ipython==8.37.0
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
joblib==1.3.2
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
kubernetes==34.1.0
langchain==0.1.13
langchain-community==0.0.34
langchain-core==0.1.53
langchain-openai==0.3.34
langchain-text-splitters==0.0.2
langcodes==3.5.0
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.1.147
language_data==1.3.0
lark==1.2.2
lxml==6.0.2
Mako==1.3.10
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.4
mlflow==3.5.1
mlflow-skinny==3.5.1
mlflow-tracing==3.5.1
mmh3==5.2.0
mpmath==1.3.0
msgspec==0.19.0
multidict==6.6.4
multiprocess==0.70.16
murmurhash==1.0.13
mypy_extensions==1.1.0
mysql-connector-python==9.4.0
mysqlclient==2.2.7
narwhals==2.5.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.2
notebook==7.2.0
notebook_shim==0.2.4
numpy==2.2.6
oauthlib==3.3.1
ollama==0.1.8
onnxruntime==1.23.2
openai==1.108.2
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
optuna==4.5.0
orjson==3.11.3
ormsgpack==1.10.0
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.5
pathvalidate==3.3.1
patsy==1.0.2
peft==0.17.1
Pillow==9.5.0
platformdirs==4.4.0
pluggy==1.6.0
posthog==6.7.11
preshed==3.0.10
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==3.20.3
psutil==7.1.0
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
Pympler==1.1
PyMySQL==1.1.2
pyparsing==3.2.5
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
pywin32==311
pywinpty==3.0.0
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.7.0
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
semver==3.0.4
Send2Trash==1.8.3
sentencepiece==0.2.1
sentry-sdk==2.39.0
shellingham==1.5.4
shtab==1.7.2
six==1.17.0
smart_open==7.4.4
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.43
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
starlette==0.41.3
statsmodels==0.14.5
streamlit==1.51.0
sympy==1.13.1
tenacity==8.5.0
terminado==0.18.1
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.11.0
tinycss2==1.4.0
tokenizers==0.22.1
toml==0.10.2
tomli==2.2.1
toolz==1.1.0
torch==2.5.1
torchao==0.14.1
torchaudio==2.8.0+cpu
torchvision==0.20.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
treelib==1.8.0
triton-windows==3.5.0.post21
trl==0.23.0
typeguard==4.4.4
typer==0.19.2
typer-slim==0.20.0
types-python-dateutil==2.9.0.20250822
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tyro==0.9.35
tzdata==2025.2
tzlocal==5.3.1
unsloth==2025.10.11
unsloth_zoo==2025.10.12
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.38.0
validators==0.35.0
waitress==3.0.2
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.2
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.3
xformers==0.0.32.post2
xgboost==2.0.3
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.25.0
